changed LtiSettings to support multiple channels instead of just one
The UX design for the NEXT Canada course now takes advantage of both
the cohort and team custom fields to specify channels for the user
launching the RiffEdu mattermost site via LTI.

Therefore the proposed UserChannels LtiSettings property of type
custom-properties.\

The CourseType exists to support specifying a particular channel
that should be the active channel after the launch. A possible way to
accomplish this would be to use a query parameter e.g.
`mm/ltilogin?setchannel=plg` or `mm/ltilogin?setchannel=capstone` using
the CourseType example values being checked in here.